% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{chat_timespan}
\alias{chat_timespan}
\title{Compute the chat duration}
\usage{
chat_timespan(x)
}
\arguments{
\item{x}{a \code{tibble} with WhatsApp chat data}
}
\value{
a \code{Period} class indicating the duration of the chat
}
\description{
Compute the chat duration
}
