% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{clean_chat}
\alias{clean_chat}
\title{Clean up WhatsApp chat data}
\usage{
clean_chat(x, omit_flag = TRUE)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix} with WhatsApp chat data}

\item{omit_flag}{should we omit "image omitted" messages from media}
}
\value{
\code{tibble} of pre-processed WhatsApp chat data
}
\description{
Clean up WhatsApp chat data
}
