% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export WhatsApp chat data to file}
\usage{
export(x, out = "", compress = TRUE)
}
\arguments{
\item{x}{\code{tibble} of WhatsApp chat data}

\item{out}{output file}

\item{compress}{boolean indicting to gzip compress output file}
}
\description{
Export WhatsApp chat data to file
}
