% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import WhatsApp chat data}
\usage{
import(x, clean = TRUE, omit_image_messages = TRUE)
}
\arguments{
\item{x}{filename to WhatsApp exported chat data}

\item{clean}{boolean to clean and preprocess the WhatsApp chat data}

\item{omit_image_messages}{boolean indicating whether to remove "omitted images" messages}
}
\value{
a \code{tibble} of pre-processed WhatsApp chat data
}
\description{
Import WhatsApp chat data
}
