% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{messages_per_day}
\alias{messages_per_day}
\title{Compute the number of messages sent per day}
\usage{
messages_per_day(x, by_user = TRUE)
}
\arguments{
\item{x}{a \code{tibble} of WhatsApp chat data}

\item{by_user}{a boolean flag to indicate if query should be grouped by user}
}
\value{
a \code{tibble} containing the number of messages sent stratified by day
}
\description{
Compute the number of messages sent per day
}
