% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{omit_image_msgs}
\alias{omit_image_msgs}
\title{Helper function to filter for omitted images}
\usage{
omit_image_msgs(x, omit = TRUE)
}
\arguments{
\item{x}{WhatsApp chat data}

\item{omit}{boolean to whether to remove "omitted images" messages}
}
\value{
\code{data.frame} with filtered or non-filtered WhatsApp chat data
}
\description{
Helper function to filter for omitted images
}
