% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_chat}
\alias{read_chat}
\title{Import WhatsApp chat data}
\usage{
read_chat(x)
}
\arguments{
\item{x}{filename of chat data}
}
\value{
raw data.frame with chat data
}
\description{
Import WhatsApp chat data
}
